vpn-controller/internal/controller/controller.go
// internal/controller/controller.go
package controller

import (
	"errors"
)

// VPNInstance represents a VPN instance.
type VPNInstance struct {
	UserID string
	IP     string
	Status string
}

// Controller manages VPN instances.
type Controller struct{}

// NewController creates a new Controller instance.
func NewController() *Controller {
	return &Controller{}
}

// CreateVPNInstance creates a new VPN instance for a given user.
func (c *Controller) CreateVPNInstance(userID string) (*VPNInstance, error) {
	if userID == "" {
		return nil, errors.New("userID cannot be empty")
	}

	// Simulate assigning an IP address
	ip := generateIP()

	// Create and return a new VPN instance
	vpn := &VPNInstance{
		UserID: userID,
		IP:     ip,
		Status: "active",
	}

	return vpn, nil
}

func (c *Controller) DeleteVPNInstance(userID string) error {
	// In a real implementation, this would involve actual deletion logic
	if userID == "" {
		return errors.New("userID cannot be empty")
	}

	// Assume deletion is successful
	return nil
}

// generateIP simulates generating an IP address for a VPN instance.
func generateIP() string {
	// This is a dummy IP generator for the sake of the example.
	// In a real-world scenario, this would interact with a cloud provider.
	return "192.168.1.1"
}
